Don't use Float32 in printing

+++ test/fixed.jl
@@ -103,7 +103,7 @@ end
 
 # Floating-point conversions
 @test isa(float(one(Fixed{Int8,6})),   Float32)
-@test isa(float(one(Fixed{Int32,18})), Float32)
+@test isa(float(one(Fixed{Int32,18})), Float64)
 @test isa(float(one(Fixed{Int32,25})), Float64)
 
 # Show
